@@ -40,7 +40,7 @@
     "css-loader": "^3.1.0",
     "html-webpack-plugin": "^3.2.0",
     "style-loader": "^0.23.1",
-    "webpack": "^4.36.1",
+    "webpack": "^4.46.0",
     "webpack-cli": "^3.3.6",
     "webpack-dev-server": "^3.7.2"
   }
