@@ -2,13 +2,13 @@ import React from 'react'
 import { FaUserFriends, FaFighterJet, FaTrophy, FaTimesCircle } from 'react-icons/fa'
 import PropTypes from 'prop-types'
 import Results from './Results'
-import { ThemeConsumer } from '../contexts/theme'
+import ThemeContext from '../contexts/theme'
 import { Link } from 'react-router-dom'
 
 function Instructions () {
+const theme = React.useContext(ThemeContext)
+
   return (
-    <ThemeConsumer>
-      {({ theme }) => (
         <div className='instructions-container'>
           <h1 className='center-text header-lg'>
             Instructions
@@ -28,32 +28,24 @@ function Instructions () {
             </li>
           </ol>
         </div>
-      )}
-    </ThemeConsumer>
-  )
+    )
 }
 
-class PlayerInput extends React.Component {
-  state = {
-    username: ''
-  }
-  handleSubmit = (event) => {
-    event.preventDefault()
+function PlayerInput ({ onSubmit, label }) {
+    const [username, setUsername] = React.useState('')
 
-    this.props.onSubmit(this.state.username)
-  }
-  handleChange = (event) => {
-    this.setState({
-      username: event.target.value
-    })
-  }
-  render() {
+    const handleSubmit = (e) => {
+      e.preventDefault()
+      onSubmit(username)    
+    }
+
+    const handleChange = (event) => setUsername(event.target.value)
+    const theme = React.useContext(ThemeContext)
+
     return (
-      <ThemeConsumer>
-        {({ theme }) => (
-          <form className='column player' onSubmit={this.handleSubmit}>
+            <form className='column player' onSubmit={handleSubmit}>
             <label htmlFor='username' className='player-label'>
-              {this.props.label}
+              {label}
             </label>
             <div className='row player-inputs'>
               <input
@@ -62,23 +54,22 @@ class PlayerInput extends React.Component {
                 className={`input-${theme}`}
                 placeholder='github username'
                 autoComplete='off'
-                value={this.state.username}
-                onChange={this.handleChange}
+                value={username}
+                onChange={handleChange}
               />
               <button
                 className={`btn ${theme === 'dark' ? 'light-btn' : 'dark-btn'}`}
                 type='submit'
-                disabled={!this.state.username}
+                disabled={!username}
               >
                 Submit
               </button>
             </div>
           </form>
-        )}
-      </ThemeConsumer>
     )
-  }
 }
+
+
 
 PlayerInput.propTypes = {
   onSubmit: PropTypes.func.isRequired,
@@ -86,9 +77,10 @@ PlayerInput.propTypes = {
 }
 
 function PlayerPreview ({ username, onReset, label }) {
-  return (
-    <ThemeConsumer>
-      {({ theme }) => (
+  const theme = React.useContext(ThemeContext)
+
+
+  return (   
         <div className='column player'>
           <h3 className='player-label'>{label}</h3>
           <div className={`row bg-${theme}`}>
@@ -109,8 +101,6 @@ function PlayerPreview ({ username, onReset, label }) {
             </button>
           </div>
         </div>
-      )}
-    </ThemeConsumer>
   )
 }
 
@@ -120,70 +110,66 @@ PlayerPreview.propTypes = {
   label: PropTypes.string.isRequired
 }
 
-export default class Battle extends React.Component {
-  state = {
-    playerOne: null,
-    playerTwo: null,
-  }
-  handleSubmit = (id, player) => {
-    this.setState({
-      [id]: player
-    })
-  }
-  handleReset = (id) => {
-    this.setState({
-      [id]: null
-    })
-  }
-  render() {
-    const { playerOne, playerTwo } = this.state
+export default function Battle () {
+  const [playerOne, setPlayerOne] = React.useState(null)
+  const [playerTwo, setPlayerTwo] = React.useState(null)
+
+  const handleSubmit = (id, player) => id === 'playerOne'
+    ? setPlayerOne(player)
+    : setPlayerTwo(player)
+
+  const handleReset = (id) => id === 'playerOne'
+  ? setPlayerOne(null)
+  : setPlayerTwo(null)  
 
     return (
-      <React.Fragment>
-        <Instructions />
+    <React.Fragment>
+      <Instructions />
 
-        <div className='players-container'>
-          <h1 className='center-text header-lg'>Players</h1>
-          <div className='row space-around'>
-            {playerOne === null
-              ? <PlayerInput
-                  label='Player One'
-                  onSubmit={(player) => this.handleSubmit('playerOne', player)}
-                />
-              : <PlayerPreview
-                  username={playerOne}
-                  label='Player One'
-                  onReset={() => this.handleReset('playerOne')}
-                />
-            }
+      <div className='players-container'>
+        <h1 className='center-text header-lg'>Players</h1>
+        <div className='row space-around'>
+          {playerOne === null
+            ? <PlayerInput
+                label='Player One'
+                onSubmit={(player) => handleSubmit('playerOne', player)}
+              />
+            : <PlayerPreview
+                username={playerOne}
+                label='Player One'
+                onReset={() => handleReset('playerOne')}
+              />
+          }
 
-            {playerTwo === null
-              ? <PlayerInput
-                  label='Player Two'
-                  onSubmit={(player) => this.handleSubmit('playerTwo', player)}
-                />
-              : <PlayerPreview
-                  username={playerTwo}
-                  label='Player Two'
-                  onReset={() => this.handleReset('playerTwo')}
-                />
-            }
-          </div>
-
-
-          {playerOne && playerTwo && (
-            <Link
-              className='btn dark-btn btn-space'
-              to={{
-                pathname: '/battle/results',
-                search: `?playerOne=${playerOne}&playerTwo=${playerTwo}`
-              }}
-            >
-              Battle
-            </Link>
-          )}
+          {playerTwo === null
+            ? <PlayerInput
+                label='Player Two'
+                onSubmit={(player) => handleSubmit('playerTwo', player)}
+              />
+            : <PlayerPreview
+                username={playerTwo}
+                label='Player Two'
+                onReset={() => handleReset('playerTwo')}
+              />
+          }
         </div>
-      </React.Fragment>
-    )
-  }
+
+
+        {playerOne && playerTwo && (
+          <Link
+            className='btn dark-btn btn-space'
+            to={{
+              pathname: '/battle/results',
+              search: `?playerOne=${playerOne}&playerTwo=${playerTwo}`
+            }}
+          >
+            Battle
+          </Link>
+        )}
+      </div>
+    </React.Fragment>
+  )
+
 }
+
+
