@@ -1,12 +1,13 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { ThemeConsumer } from '../contexts/theme'
+import ThemeContext from '../contexts/theme'
 
 export default function Card ({ header, subheader, avatar, href, name, children }) {
+ const theme = React.useContext(ThemeContext)
+ 
+ 
   return (
-    <ThemeConsumer>
-      {({ theme }) => (
-        <div className={`card bg-${theme}`}>
+       <div className={`card bg-${theme}`}>
           <h4 className='header-lg center-text'>
             {header}
           </h4>
@@ -27,9 +28,7 @@ export default function Card ({ header, subheader, avatar, href, name, children 
           </h2>
           {children}
         </div>
-      )}
-    </ThemeConsumer>
-  )
+   )
 }
 
 Card.propTypes = {
