@@ -1,6 +1,7 @@
 import React from 'react'
 
-const { Consumer, Provider } = React.createContext()
+const ThemeContext = React.createContext()
 
-export const ThemeConsumer = Consumer
-export const ThemeProvider = Provider
+export default ThemeContext
+export const ThemeConsumer = ThemeContext.Consumer
+export const ThemeProvider = ThemeContext.Provider 
